move logic to manager and repository classes
