+++ server/src/main/kotlin/com/github/csc3380fall2024/team16/plugins/Rpc.kt
@@ -1,11 +1,9 @@
 package com.github.csc3380fall2024.team16.plugins
 
 import com.github.csc3380fall2024.team16.RpcService
-import com.github.csc3380fall2024.team16.createHash
-import com.github.csc3380fall2024.team16.createRandomSalt
-import com.github.csc3380fall2024.team16.createToken
+import com.github.csc3380fall2024.team16.TokenManager
 import com.github.csc3380fall2024.team16.emailRegex
-import com.github.csc3380fall2024.team16.userFromToken
+import com.github.csc3380fall2024.team16.repository.UserRepository
 import com.github.csc3380fall2024.team16.usernameRegex
 import io.ktor.server.application.Application
 import io.ktor.server.application.install
@@ -13,13 +11,6 @@ import io.ktor.server.routing.routing
 import kotlinx.rpc.krpc.ktor.server.RPC
 import kotlinx.rpc.krpc.ktor.server.rpc
 import kotlinx.rpc.krpc.serialization.json.json
-import org.jetbrains.exposed.sql.SqlExpressionBuilder.eq
-import org.jetbrains.exposed.sql.Table
-import org.jetbrains.exposed.sql.exists
-import org.jetbrains.exposed.sql.insertAndGetId
-import org.jetbrains.exposed.sql.or
-import org.jetbrains.exposed.sql.selectAll
-import org.jetbrains.exposed.sql.transactions.transaction
 import kotlin.coroutines.CoroutineContext
 
 fun Application.configureRpc() {
@@ -47,30 +38,8 @@ class RpcServiceImpl(override val coroutineContext: CoroutineContext) : RpcServi
             password.isEmpty()               -> throw Exception("password must not be empty")
         }
         
-        val salt = createRandomSalt()
-        val hash = createHash(password, salt)
-        val id = transaction {
-            val usernameExists = run {
-                val existsOp = exists(Users.selectAll().where(Users.username eq username))
-                Table.Dual.select(existsOp).first()[existsOp]
-            }
-            if (usernameExists) throw Exception("user with that username already exists")
-            
-            val emailExists = run {
-                val existsOp = exists(Users.selectAll().where(Users.email eq email))
-                Table.Dual.select(existsOp).first()[existsOp]
-            }
-            if (emailExists) throw Exception("user with that email already exists")
-            
-            Users.insertAndGetId {
-                it[Users.username] = username
-                it[Users.email] = email
-                it[passwordSalt] = salt
-                it[passwordHash] = hash
-            }
-        }
-        
-        return createToken(id.value, hash)
+        val user = UserRepository.addUser(username, email, password)
+        return TokenManager.createToken(user.id, user.passwordHash)
     }
     
     override suspend fun login(usernameOrEmail: String, password: String): String {
@@ -79,17 +48,7 @@ class RpcServiceImpl(override val coroutineContext: CoroutineContext) : RpcServi
             password.isEmpty()        -> throw Exception("password must not be empty")
         }
         
-        val row = transaction {
-            Users.selectAll()
-                .where { (Users.username eq usernameOrEmail) or (Users.email eq usernameOrEmail) }
-                .limit(1)
-                .firstOrNull() ?: throw Exception("invalid credentials")
-        }
-        
-        if (!createHash(password, row[Users.passwordSalt]).contentEquals(row[Users.passwordHash])) {
-            throw Exception("invalid credentials")
-        }
-        
-        return createToken(row[Users.id].value, row[Users.passwordHash])
+        val user = UserRepository.getUser(usernameOrEmail, password) ?: throw Exception("invalid credentials")
+        return TokenManager.createToken(user.id, user.passwordHash)
     }
 }
